# packages/api-client/openapi/reboxe.yaml
openapi: 3.1.0
info:
  title: REBOXE API
  version: 1.4.0
servers:
  - url: http://127.0.0.1:3001
paths:
...
  /v1/admin/kyc:
    get:
      operationId: listKycReviews
      description: Requires database staff role MODERATOR or SUPER_ADMIN, ACTIVE profile/role, and verified JWT aal2.
      security:
      - {bearerAuth: []}
      parameters:
      - name: status
        in: query
        schema: {type: string, enum: [MANUAL_REVIEW], default: MANUAL_REVIEW}
      - name: cursor
        in: query
        schema: {type: string, maxLength: 200}
        description: Opaque nextCursor from the previous page.
      responses:
        '200':
          description: Authorized KYC review response
          headers:
            Cache-Control:
              schema: {type: string, const: no-store}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminKycQueue'}
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '403':
          description: Active moderator/super admin and AAL2 required
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '404':
          description: KYC not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '422':
          description: Invalid query, body or key
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
  /v1/admin/kyc/{kycId}:
    get:
      operationId: getKycReview
      description: Requires database staff role MODERATOR or SUPER_ADMIN, ACTIVE profile/role, and verified JWT aal2.
      security:
      - {bearerAuth: []}
      parameters:
      - name: kycId
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: Authorized KYC review response
          headers:
            Cache-Control:
              schema: {type: string, const: no-store}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminKycDetail'}
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '403':
          description: Active moderator/super admin and AAL2 required
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '404':
          description: KYC not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '422':
          description: Invalid query, body or key
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
  /v1/admin/kyc/{kycId}/decision:
    post:
      operationId: decideKycReview
      description: Requires database staff role MODERATOR or SUPER_ADMIN, ACTIVE profile/role, and verified JWT aal2.
      security:
      - {bearerAuth: []}
      parameters:
      - name: kycId
        in: path
        required: true
        schema: {type: string}
      - name: Idempotency-Key
        in: header
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200':
          description: Authorized KYC review response
          headers:
            Cache-Control:
              schema: {type: string, const: no-store}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/KycDecisionResult'}
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '403':
          description: Active moderator/super admin and AAL2 required
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '404':
          description: KYC not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '422':
          description: Invalid query, body or key
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '409':
          description: Already decided or idempotency key reused with a different payload
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/KycDecision'}
...
  /v1/checkout/init:
    post:
      operationId: initSandboxCheckout
      security: [{ bearerAuth: [] }]
      parameters: [{ in: header, name: Idempotency-Key, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CheckoutInit" } } }
      responses: { "201": { description: Reserved canonical sandbox order } }
  /v1/checkout/{orderId}/pay:
    post:
      operationId: confirmSandboxCod
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: orderId, required: true, schema: { type: string } }
        - { in: header, name: Idempotency-Key, required: true, schema: { type: string } }
      responses: { "201": { description: Confirmed SANDBOX_COD order; no real payment } }
  /v1/orders:
    get:
      operationId: listCanonicalBuyerOrders
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Buyer canonical orders } }
  /v1/orders/{orderId}:
    get:
      operationId: getCanonicalBuyerOrder
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: orderId, required: true, schema: { type: string } }]
      responses: { "200": { description: Buyer order detail and timeline } }
  /v1/shops/{shopId}/orders:
    get:
      operationId: listCanonicalSellerOrders
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: shopId, required: true, schema: { type: string } }]
      responses: { "200": { description: Seller canonical orders } }
  /v1/shops/{shopId}/finance/projection:
    get:
      operationId: getSyntheticFinanceProjection
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: shopId, required: true, schema: { type: string } }]
      responses: { "200": { description: Ledger-derived synthetic buckets } }
  /v1/shops/{shopId}/orders/{orderId}/shipment:
    post:
      operationId: createFakeShipment
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: shopId, required: true, schema: { type: string } }
        - { in: path, name: orderId, required: true, schema: { type: string } }
      responses: { "201": { description: Deterministic fake shipment and synthetic label } }
  /v1/orders/{orderId}/disputes:
    post:
      operationId: openDispute
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: orderId, required: true, schema: { type: string } }]
      responses: { "201": { description: Dispute accepted without mandatory evidence } }
  /v1/processing-records:
    post:
      operationId: createProcessingRecord
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Versioned processing notice record } }
  /v1/disputes/{caseId}/evidence:
    get:
      operationId: listAuthorizedEvidence
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: caseId, required: true, schema: { type: string } }]
      responses: { "200": { description: Seller receives derivatives only } }
    post:
      operationId: addFakeEvidence
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: caseId, required: true, schema: { type: string } }]
      responses: { "201": { description: Versioned fake evidence metadata } }
  /v1/disputes/{caseId}/replies:
    post:
      operationId: replyDispute
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: caseId, required: true, schema: { type: string } }]
      responses: { "201": { description: Audited buyer or seller response } }
  /v1/disputes/{caseId}/appeal:
    post:
      operationId: appealDispute
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: caseId, required: true, schema: { type: string } }]
      responses: { "201": { description: Appeal accepted inside the appeal window } }
  /v1/admin/disputes:
    get:
      operationId: listAdminDisputes
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: AAL2 dispute queue } }
  /v1/admin/disputes/{caseId}/decision:
    post:
      operationId: resolveDispute
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: caseId, required: true, schema: { type: string } }]
      responses: { "201": { description: Refund obligation, never a paid payout } }
  /v1/notifications:
    get:
      operationId: listInAppNotifications
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: In-app notification inbox } }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: {type: string}
            message: {type: string}
            requestId: {type: string}
...
    KycReview:
      type: object
      additionalProperties: false
      required: [reason, reviewedAt]
      properties:
        reason: {type: string}
        reviewedAt: {type: string, format: date-time}
    KycIdentity:
      type: object
      additionalProperties: false
      required: [citizenId, fullName, dateOfBirth, gender, address, issuedAt]
      properties:
        citizenId: {type: [string, 'null'], description: Masked identifier; only the last four digits are shown.}
        fullName: {type: [string, 'null']}
        dateOfBirth: {type: [string, 'null']}
        gender: {type: [string, 'null']}
        address: {type: [string, 'null']}
        issuedAt: {type: [string, 'null']}
    KycVerification:
      type: object
      additionalProperties: false
      required: [documentValid, faceMatched, faceScore, livenessPassed, livenessScore]
      properties:
        documentValid: {type: [boolean, 'null']}
        faceMatched: {type: [boolean, 'null']}
        faceScore: {type: [number, 'null']}
        livenessPassed: {type: [boolean, 'null']}
        livenessScore: {type: [number, 'null']}
    AdminKycQueueItem:
      type: object
      additionalProperties: false
      required: [kycId, shopId, shopDisplayName, status, provider, submittedAt]
      properties:
        kycId: {type: string}
        shopId: {type: string}
        shopDisplayName: {type: string}
        status: {type: string, const: MANUAL_REVIEW}
        provider: {type: string}
        submittedAt: {type: string, format: date-time}
    AdminKycQueue:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/AdminKycQueueItem'}
        nextCursor: {type: [string, 'null']}
    KycDecision:
      type: object
      additionalProperties: false
      required: [decision, reason]
      properties:
        decision: {type: string, enum: [APPROVE, REJECT]}
        reason: {type: string, minLength: 1, maxLength: 1000, description: Trimmed non-empty reason visible to seller.}
    KycDecisionResult:
      type: object
      additionalProperties: false
      required: [kycId, kycStatus, review]
      properties:
        kycId: {type: string}
        kycStatus: {type: string, enum: [VERIFIED, REJECTED]}
        review: {$ref: '#/components/schemas/KycReview'}
    AdminKycDetail:
      type: object
      additionalProperties: false
      required: [kycId, shopId, shopDisplayName, status, provider, submittedAt, identity, verification, review, tax, bank]
      properties:
        kycId: {type: string}
        shopId: {type: string}
        shopDisplayName: {type: string}
        status: {type: string, enum: [PENDING, PROCESSING, VERIFIED, REJECTED, MANUAL_REVIEW]}
        provider: {type: string}
        submittedAt: {type: string, format: date-time}
        identity: {$ref: '#/components/schemas/KycIdentity'}
        verification: {$ref: '#/components/schemas/KycVerification'}
        review:
          anyOf:
          - {$ref: '#/components/schemas/KycReview'}
          - {type: 'null'}
        tax:
          type: object
          additionalProperties: false
          required: [status, registeredName]
          properties:
            status:
              anyOf:
              - {type: string, enum: [VERIFIED, NOT_FOUND, UNAVAILABLE]}
              - {type: 'null'}
            registeredName: {type: [string, 'null']}
        bank:
          type: object
          additionalProperties: false
          required: [bankCode, accountNumber, status, registeredName, nameMatchScore]
          properties:
            bankCode: {type: [string, 'null']}
            accountNumber: {type: [string, 'null'], description: Masked account number.}
            status:
              anyOf:
              - {type: string, enum: [VERIFIED, NOT_FOUND, UNAVAILABLE]}
              - {type: 'null'}
            registeredName: {type: [string, 'null']}
            nameMatchScore: {type: [number, 'null']}
...
    CheckoutInit:
      type: object
      additionalProperties: false
      required: [items, addressId]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [listingId, quantity]
            properties:
              listingId: { type: string }
              quantity: { type: integer, enum: [1] }
        addressId: { type: string }
...